# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/departments": {
      "post": {
        "operationId": "DepartmentsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateDepartmentDto" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Department created successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Department" }
              }
            }
          },
          "400": { "description": "Bad request" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" }
        },
        "security": [{ "bearer": [] }, { "bearer": [] }],
        "summary": "Create a new department",
        "tags": ["Departments"]
      },
      "get": {
        "operationId": "DepartmentsController_findAll",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "query",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "List of departments",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Department" }
                }
              }
            }
          },
          "401": { "description": "Unauthorized" }
        },
        "security": [{ "bearer": [] }],
        "summary": "Get all departments for an organization",
        "tags": ["Departments"]
      }
    },
    "/api/departments/hierarchy": {
      "get": {
        "operationId": "DepartmentsController_getHierarchy",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "query",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Hierarchical list of departments",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Department" }
                }
              }
            }
          },
          "401": { "description": "Unauthorized" }
        },
        "security": [{ "bearer": [] }],
        "summary": "Get department hierarchy for an organization",
        "tags": ["Departments"]
      }
    },
    "/api/departments/with-users": {
      "get": {
        "operationId": "DepartmentsController_findAllWithUsers",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "query",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "List of departments with users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Department" }
                }
              }
            }
          },
          "401": { "description": "Unauthorized" }
        },
        "security": [{ "bearer": [] }],
        "summary": "Get all departments with users for an organization",
        "tags": ["Departments"]
      }
    },
    "/api/departments/{id}": {
      "get": {
        "operationId": "DepartmentsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          },
          {
            "name": "organizationId",
            "required": true,
            "in": "query",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Department found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Department" }
              }
            }
          },
          "401": { "description": "Unauthorized" },
          "404": { "description": "Department not found" }
        },
        "security": [{ "bearer": [] }],
        "summary": "Get a department by ID",
        "tags": ["Departments"]
      },
      "put": {
        "operationId": "DepartmentsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          },
          {
            "name": "organizationId",
            "required": true,
            "in": "query",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateDepartmentDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Department updated successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Department" }
              }
            }
          },
          "400": { "description": "Bad request" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" },
          "404": { "description": "Department not found" }
        },
        "security": [{ "bearer": [] }, { "bearer": [] }],
        "summary": "Update a department",
        "tags": ["Departments"]
      },
      "delete": {
        "operationId": "DepartmentsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          },
          {
            "name": "organizationId",
            "required": true,
            "in": "query",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Department deleted successfully" },
          "400": { "description": "Bad request" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" },
          "404": { "description": "Department not found" }
        },
        "security": [{ "bearer": [] }, { "bearer": [] }],
        "summary": "Delete a department",
        "tags": ["Departments"]
      }
    },
    "/api/departments/by-slug/{slug}": {
      "get": {
        "operationId": "DepartmentsController_findBySlug",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          },
          {
            "name": "organizationId",
            "required": true,
            "in": "query",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Department found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Department" }
              }
            }
          },
          "401": { "description": "Unauthorized" },
          "404": { "description": "Department not found" }
        },
        "security": [{ "bearer": [] }],
        "summary": "Get a department by slug",
        "tags": ["Departments"]
      }
    },
...
  },
  "info": {
    "title": "eTickets Platform API",
    "description": "The eTickets Platform API documentation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": { "scheme": "bearer", "bearerFormat": "JWT", "type": "http" }
    },
    "schemas": {
...
      "DepartmentNotificationPreferencesDto": {
        "type": "object",
        "properties": {
          "newTicket": {
            "type": "boolean",
            "description": "Notify on new tickets"
          },
          "ticketAssigned": {
            "type": "boolean",
            "description": "Notify on ticket assignments"
          },
          "ticketClosed": {
            "type": "boolean",
            "description": "Notify on ticket closures"
          },
          "dailySummary": {
            "type": "boolean",
            "description": "Send daily summary"
          }
        }
      },
      "DepartmentCustomFieldDto": {
        "type": "object",
        "properties": {
          "key": { "type": "string", "description": "Field key" },
          "label": { "type": "string", "description": "Field label" },
          "type": {
            "type": "string",
            "description": "Field type",
            "enum": ["text", "number", "date", "boolean", "select"]
          },
          "options": {
            "description": "Field options for select type",
            "type": "array",
            "items": { "type": "string" }
          },
          "required": {
            "type": "boolean",
            "description": "Whether the field is required"
          }
        },
        "required": ["key", "label", "type"]
      },
      "DepartmentSettingsDto": {
        "type": "object",
        "properties": {
          "color": { "type": "string", "description": "Department color" },
          "icon": { "type": "string", "description": "Department icon" },
          "contactEmail": { "type": "string", "description": "Contact email" },
          "contactPhone": { "type": "string", "description": "Contact phone" },
          "location": {
            "type": "string",
            "description": "Department location"
          },
          "defaultTicketAssignee": {
            "type": "string",
            "description": "Default ticket assignee ID"
          },
          "autoAssignTickets": {
            "type": "boolean",
            "description": "Auto-assign tickets"
          },
          "notificationPreferences": {
            "description": "Notification preferences",
            "allOf": [
              {
                "$ref": "#/components/schemas/DepartmentNotificationPreferencesDto"
              }
            ]
          },
          "customFields": {
            "description": "Custom fields",
            "type": "array",
            "items": { "$ref": "#/components/schemas/DepartmentCustomFieldDto" }
          }
        }
      },
      "CreateDepartmentDto": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "description": "Department name" },
          "description": {
            "type": "string",
            "description": "Department description"
          },
          "slug": {
            "type": "string",
            "description": "URL-friendly identifier"
          },
          "organizationId": {
            "type": "string",
            "description": "Organization ID"
          },
          "headId": {
            "type": "string",
            "description": "Department head user ID"
          },
          "parentDepartmentId": {
            "type": "string",
            "description": "Parent department ID"
          },
          "code": {
            "type": "string",
            "description": "Department code or identifier"
          },
          "settings": {
            "description": "Department settings",
            "allOf": [{ "$ref": "#/components/schemas/DepartmentSettingsDto" }]
          },
          "isActive": {
            "type": "boolean",
            "description": "Whether the department is active"
          }
        },
        "required": ["name", "slug", "organizationId"]
      },
      "Department": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier for the department"
          },
          "name": { "type": "string", "description": "Department name" },
          "description": {
            "type": "string",
            "description": "Department description"
          },
          "slug": {
            "type": "string",
            "description": "URL-friendly identifier"
          },
          "organizationId": {
            "type": "string",
            "description": "ID of the organization the department belongs to"
          },
          "headId": {
            "type": "string",
            "description": "Department head user ID"
          },
          "parentDepartmentId": {
            "type": "string",
            "description": "Parent department ID"
          },
          "code": {
            "type": "string",
            "description": "Department code or identifier"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "When the department was created"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "When the department was last updated"
          },
          "settings": {
            "type": "object",
            "description": "Department settings and metadata"
          },
          "isActive": {
            "type": "boolean",
            "description": "Whether the department is active"
          }
        },
        "required": [
          "id",
          "name",
          "slug",
          "organizationId",
          "createdAt",
          "updatedAt",
          "settings",
          "isActive"
        ]
      },
      "UpdateDepartmentDto": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "description": "Department name" },
          "description": {
            "type": "string",
            "description": "Department description"
          },
          "slug": {
            "type": "string",
            "description": "URL-friendly identifier"
          },
          "organizationId": {
            "type": "string",
            "description": "Organization ID"
          },
          "headId": {
            "type": "string",
            "description": "Department head user ID"
          },
          "parentDepartmentId": {
            "type": "string",
            "description": "Parent department ID"
          },
          "code": {
            "type": "string",
            "description": "Department code or identifier"
          },
          "settings": {
            "description": "Department settings",
            "allOf": [{ "$ref": "#/components/schemas/DepartmentSettingsDto" }]
          },
          "isActive": {
            "type": "boolean",
            "description": "Whether the department is active"
          }
        }
      },
...
    }
  }
}